+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * 【core   】     DatePattern增加年月格式化常量（pr#220@Gitee）
 * 【core   】     ArrayUtil增加shuffle方法（pr#1255@Github）
 * 【core   】     ArrayUtil部分方法分离至PrimitiveArrayUtil
+* 【crypto 】     opt改为otp包（issue#1257@Github）
 
 ### Bug修复
 * 【cron   】     修复CronTimer可能死循环的问题（issue#1224@Github）
