+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/otp/HOTP.java
@@ -1,4 +1,4 @@
-package cn.hutool.crypto.digest.opt;
+package cn.hutool.crypto.digest.otp;
 
 import cn.hutool.crypto.digest.HMac;
 import cn.hutool.crypto.digest.HmacAlgorithm;
